Update activity
Change activity name: MainActivity -> LoginActivity.
