+++ app/src/main/java/com/uof/uof_mobile/LoginActivity.java
@@ -1,0 +1,14 @@
+package com.uof.uof_mobile;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.os.Bundle;
+
+public class LoginActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_login);
+    }
+}
